Add null checking
I've decided to add extensive null checking to all structures. null is
not a functional way to deal with absence of values. There's Option for
that. So now almost every function throws ArgumentNullException if it
gets a null as a parameter.

As before, if an option receives a null, it becomes empty. Lists simply
ignore nulls, so a list cannot contain them at all.