Add UnacceptableNullException
As it turns out, there're no exception types for nulls, apart from
NullReferenceException (thrown when trying to dereference a null value)
and ArgumentNullException (thrown when an argument is null). So I added
a general "null is not welcome" exception.

Map and Bind/FlatMap in the data structures now throw this exception when
a mapper function returns a null.

Also, added some tests for Some (I discovered this flaw when writing
tests).

In addition, now the exceptions are in a dedicated Exceptions namespace.