Make ConsList<T> implement collection interfaces
Now ConsList<T> is not just an IEnumeratble, it's also an ICollection
and an IList (both generic and non-generic).

I've decided that it would be better to implement those interfaces for
interoperation with the usual C# ways.

Since cons lists are immutable, many methods of those interfaces throw
an exception. Some methods are not optimized.